% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od_any_opioid_icd10cm.R
\name{add_any_opioid_icd10cm}
\alias{add_any_opioid_icd10cm}
\title{Any opioid case.}
\usage{
add_any_opioid_icd10cm(data, diag_ecode_col)
}
\arguments{
\item{data:}{input data}

\item{diag_ecode_col:}{indices of diagnosis and
E codes of interest}
}
\value{
variable any_opioid_icd10cm
}
\description{
Any opioid case.
}
\examples{
to be added
}
