% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od_non_heroin_icd10cm.R
\name{add_non_heroin_icd10cm}
\alias{add_non_heroin_icd10cm}
\title{Title}
\usage{
add_non_heroin_icd10cm(data, diag_ecode_col)
}
\arguments{
\item{diag_ecode_col}{}
}
\value{
non_heroin_icd10cm
}
\description{
Title
}
