% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_diag.R
\name{create_diag}
\alias{create_diag}
\title{Create a new variable based on pattern.}
\usage{
create_diag(data, expr, colvec, ignore.case = T, perl = T)
}
\arguments{
\item{ignore.case}{logical}

\item{perl}{logical}

\item{data:}{input data}

\item{expr:}{regular expression describing the pattern of interest}

\item{colvec:}{indices of variables of interest}
}
\value{
new variable matching the pattern described in the regular expression
}
\description{
Create a new variable based on pattern in the argument expr
}
\examples{

}
