% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dea_set.R
\docType{data}
\name{dea_set}
\alias{dea_set}
\title{Simulated death dataset of 6 variables.}
\format{A data frame with 500 rows and 6 variables:
\describe{
  \item{age}{age, in years}
  \item{year}{year of death}
  \item{ucod}{underlying cause}
  \item{cause_1}{multiple cause 1}
  \item{cause_2}{multiple cause 2}
  \item{cause_3}{multiple cause 3}
}}
\usage{
dea_set
}
\description{
A dataset containing underlying cause of death
 and other attributes of 500 deaths
}
\keyword{datasets}
