% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{format_multirow_ems_data}
\alias{format_multirow_ems_data}
\title{Converts multi-row data to single row data}
\usage{
format_multirow_ems_data(
  data_in,
  event_id_name = "CAD Incident Number (eResponse.03)",
  patient_first_name = "Patient First Name (ePatient.03)",
  patient_last_name = "Patient Last Name (ePatient.02)",
  patient_age_name = "Patient Age (ePatient.15)",
  primary_impression_name = "Situation Provider Primary Impression (eSituation.11)",
  primary_complaint_name = "Situation Primary Complaint Statement List (eSituation.04)",
  care_narrative_name = "Patient Care Report Narrative (eNarrative.01)",
  medication_given_name = "Medication Given Description (eMedications.03)",
  medication_response_name = "Medication Response (eMedications.07)",
  remove_numbers_from_text = TRUE,
  stop_words = overdoseR:::stop_words
)
}
\arguments{
\item{data_in}{raw EMS data to use}

\item{event_id_name}{Name of id column, defaults to "CAD Incident Number (eResponse.03)",}

\item{patient_first_name}{Name of first name column, defaults to "Patient First Name (ePatient.03)",}

\item{patient_last_name}{Name of last name column, defaults to (ePatient.02)",}

\item{patient_age_name}{Name of age column, defaults to (ePatient.15)",}

\item{primary_impression_name}{Name of primary impression column, defaults to "Situation Provider Primary Impression (eSituation.11)",}

\item{primary_complaint_name}{Name of primary complaint column, defaults to "Situation Primary Complaint Statement List (eSituation.04)",}

\item{care_narrative_name}{Name of care narrative column, defaults to "Patient Care Report Narrative (eNarrative.01)",}

\item{medication_given_name}{Name of medication administered column, defaults to "Medication Given Description (eMedications.03)",}

\item{medication_response_name}{Name of medication response column, defaults to "Medication Response (eMedications.07)",}

\item{remove_numbers_from_text}{Logical flag to remove numbers from text}

\item{stop_words}{Corpus of stop words, defaults to tidytext::stop_words (tidytext not required)}
}
\value{
One row per record dataframe of inputted data and one-hot fields
}
\description{
Converts multi-row data to single row data
}
\examples{
\dontrun{format_multirow_ems_data(data_in = raw_ems_data)}
}
