% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{high_value_terms}
\alias{high_value_terms}
\title{High Value Terms}
\format{A data frame with 189 rows and 4 variables:
\describe{
  \item{token}{actual string of high value term}
  \item{freqYES}{Frequency of term in development set for opioid related events}
  \item{freqNO}{Frequency of term in development set for non-opioid related events}
  \item{type}{word, bigram, or trigram}
}}
\usage{
high_value_terms
}
\description{
Dataset containing the raw words, bigrams, and trigrams that
were determined to be statistically significant in prior research.
}
\keyword{datasets}
