% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hosp_set}
\alias{hosp_set}
\title{Simulated Hospitalization Data.}
\format{Data frame}
\source{
simulation.
}
\usage{
hosp_set
}
\description{
Dataset of 500 rows and 8 variables.
}
\examples{
library(dplyr)
sample_n(hosp_set, 5)

}
\keyword{datasets}
