% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd10cm18.R
\docType{data}
\name{icd10cm18}
\alias{icd10cm18}
\title{ICD-10 CM table.}
\format{A data frame with 71704 rows and 2 variables: \describe{
  \item{icd10cm}{icd10cm codes} \item{icd10cm_description}{icd10cm code
  description} }}
\source{
\url{https://www.cms.gov/Medicare/Coding/ICD10/2018-ICD-10-CM-and-GEMs.html}
}
\usage{
icd10cm18
}
\description{
ICD-10 CM table with code and description
}
\keyword{datasets}
