% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{icd9cm14}
\alias{icd9cm14}
\title{ICD-9-CM codes table and description v2014.}
\format{Data frame}
\source{
\url{https://www.cms.gov/Medicare/Coding/ICD9ProviderDiagnosticCodes/codes.html}
}
\usage{
icd9cm14
}
\description{
Dataset of 14567 rows and 2 variables.
}
\examples{
library(dplyr)
sample_n(icd9cm14, 5)

}
\keyword{datasets}
