% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{make_one_hot}
\alias{make_one_hot}
\title{Creates one-hot-matrix of terms in narrative. Intended to be used internally with apply.}
\usage{
make_one_hot(row_data_in, word_tokens, bigram_tokens, trigram_tokens)
}
\arguments{
\item{row_data_in}{String of single narrative}

\item{word_tokens}{Words to tokenize on}

\item{bigram_tokens}{Bigrams to tokenize on}

\item{trigram_tokens}{Trigrams to tokenize on}
}
\value{
Vector of one-hot terms
}
\description{
Creates one-hot-matrix of terms in narrative. Intended to be used internally with apply.
}
