% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od_age11.R
\name{od_age11}
\alias{od_age11}
\title{Title}
\usage{
od_age11(data, age)
}
\arguments{
\item{data:}{input data}

\item{age:}{age in years}
}
\value{
new variables of 11 age groups agegrp11 and age11
}
\description{
Title
}
