% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od_any_opioid_icd10cm.R
\name{od_any_opioid_icd10cm}
\alias{od_any_opioid_icd10cm}
\title{Any opioid case.}
\usage{
od_any_opioid_icd10cm(data, diag_ecode_col, any_opioid_icd10cm)
}
\arguments{
\item{data:}{input data}

\item{diag_ecode_col:}{indices of diagnosis and
E codes of interest}

\item{any_opioid_icd10cm:}{choose the name for the created variable}
}
\value{
variable any_opioid_icd10cm (or the chosen name)
}
\description{
Any opioid case.
}
\examples{
to be added
}
