% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od_cocaine_icd10cm.R
\name{od_cocaine_icd10cm}
\alias{od_cocaine_icd10cm}
\title{Find cocain based on ICD-10-CM.}
\usage{
od_cocaine_icd10cm(data, diag_ecode_col)
}
\arguments{
\item{data}{input data}

\item{diag_ecode_col}{column indices}
}
\value{
cocaine_icd10cm
}
\description{
Find cocaine based on ICD-10-CM.
}
\examples{
to be added
}
