% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od_create_cond_diag.R
\name{od_create_cond_diag}
\alias{od_create_cond_diag}
\title{create conditionnally a new variable.}
\usage{
od_create_cond_diag(data, expr, colvec, ignore.case = T, perl = T, cond.var)
}
\arguments{
\item{ignore.case}{logical}

\item{perl}{logical}

\item{data:}{input data}

\item{expr:}{regular expressions}

\item{colvec:}{indices of variables of interest}

\item{cond.var:}{conditional variable}
}
\value{
new variable defined by the regular expressions
}
\description{
create conditionnally a new variable
}
\examples{
to be added
}
