% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od_create_diag.r
\name{od_create_diag}
\alias{od_create_diag}
\title{Title}
\usage{
od_create_diag(data, expr, colvec, ignore.case = T, perl = T)
}
\arguments{
\item{perl}{}

\item{data:}{input data}

\item{expr:}{regular expression describing
the pattern of interest}

\item{colvec:}{indices of variables of interest}
}
\value{
new variable matching the pattern
described in the regular expression
}
\description{
Title
}
\examples{
heroin_icd9cm_regex_ <- "^96501|^E8500"
library(tidyverse)
	hosp_net \%>\%
	filter(year == 2015) \%>\%
	od_create_diag(expr = heroin_icd9cm_regex_,
	colvec = c(1:5)) \%>\% sample_n(10)

}
