% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od_create_diag_9.R
\name{od_create_diag_9}
\alias{od_create_diag_9}
\title{Create a new variable based on patterns.}
\usage{
od_create_diag_9(data, expr1, colvec1, expr2, colvec2, ignore.case = T,
  perl = T)
}
\arguments{
\item{ignore.case}{logical}

\item{perl}{logical}

\item{data:}{input data}

\item{expr1:}{regular expression describing the pattern of interest}

\item{colvec1:}{indices of variables of interest}

\item{expr2:}{regular expression describing the pattern of interest}

\item{colvec2:}{indices of variables of interest}
}
\value{
new variable matching the pattern described in the regular expressions
}
\description{
Create a new variable based on pattern in the argument expr
}
\examples{
library(tidyverse)
\\thosp_set \%>\%
\\tfilter(year == 2015, quarters(discharge_date) == "Q1") \%>\%
\\tod_create_diag2(., expr1 = "^96501", colvec1 = 1,
\\texpr2 = "^E8500", colvec2 = c(1:5)) \%>\% sample_n(10)


}
