% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od_date_day.r
\name{od_date_day}
\alias{od_date_day}
\title{Complete trucated date.}
\usage{
od_date_day(x)
}
\arguments{
\item{x}{only truncated date of mm/yyyy or m/yyyy}
}
\value{
Return complete date with 15/ inserted as day
}
\description{
Complete trucated date.
}
\examples{
od_date_day(c('03/2015', '11/2014', '09/2017'))
# '2015-03-15' '2014-11-15' '2017-09-15'
}
