% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od_diag9_10cm.R
\name{od_diag9_10cm}
\alias{od_diag9_10cm}
\title{Find diagnoses across ICD-9 and ICD-10 periods.}
\usage{
od_diag9_10cm(data, diag_ecode_col, expr9, expr10, var_name, date)
}
\arguments{
\item{data}{input data}

\item{expr9}{the ICD-9 regular expressions}

\item{expr10}{the ICD-9 regular expressions}

\item{var_name}{name for the new variable}

\item{underly_col}{underlying column index}

\item{expr10}{the date column}
}
\value{
new variable named with var_name
}
\description{
Find diagnoses across ICD-9 and ICD-10 periods.
}
\examples{
to be added
}
