% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od_direct_age_adjust.R
\name{od_direct_age_adjust}
\alias{od_direct_age_adjust}
\title{Title}
\usage{
od_direct_age_adjust(data, agegrp, count,
  population, standard_pop = NULL, s = 1e+05, r = 1,
  alpha = 0.05)
}
\arguments{
\item{data}{input data}

\item{agegrp}{age group variable}

\item{count}{count variable}

\item{population}{population variable}

\item{standard_pop}{standar population table or NULL}

\item{s}{scale numeric}

\item{r}{round numeric}

\item{alpha}{alpha numeric}
}
\value{
age_adj_rate, lower_age_adj
and upper_age_adj
}
\description{
Title
}
\examples{
to be added
}
