% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od_drug_apr_icd10cm.R
\name{od_drug_apr_icd10cm}
\alias{od_drug_apr_icd10cm}
\title{Find drug types from ICD-10-CM.}
\usage{
od_drug_apr_icd10cm(data, diag_ecode_col)
}
\arguments{
\item{data}{input data}

\item{diag_ecode_col}{column indices}
}
\value{
any_drug, any_opioid, non_heroin_opioid, and
  heroin
}
\description{
Find any drug and selected opioids
}
\examples{
to be added
library(tidyverse)
filter(hosp_set, year == 2016, quarters(discharge_date) == "Q4" ) \%>\% od_drug_apr_icd10cm(diag_ecode_col = c(3, 6)) \%>\% sample_n(5)

}
