% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od_drug_apr_icd9cm.R
\name{od_drug_apr_icd9cm}
\alias{od_drug_apr_icd9cm}
\title{Find drug types based on ICD-9-CM.}
\usage{
od_drug_apr_icd9cm(data, diag_col, ecode_col)
}
\arguments{
\item{data:}{input data}

\item{diag_col:}{indices of principal diagnosis}

\item{ecode_col:}{indices of ecodes}
}
\value{
any_drug, any_opioid, non_heroin_opioid, and
  heroin
}
\description{
Find any drug and other opioids from ICD-9-CM
}
\examples{
library(tidyverse)
filter(hosp_set, year == 2014) \%>\% od_drug_apr_icd9cm(diag_col = 3, ecode_col c(3:5)) \%>\% sample_n(5)

}
