% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od_drug_apr_trans.R
\name{od_drug_apr_trans}
\alias{od_drug_apr_trans}
\title{Add new drug variables based on both ICD-9-CM and ICD-10-CM.}
\usage{
od_drug_apr_trans(data, diag_col, ecode_col, diag_ecode_col, date)
}
\arguments{
\item{data:}{the input dataset}

\item{diag_col:}{ndices(numeric) of  diag of interest for icd-9}

\item{diag_ecode:}{indices(numeric) of  ecodes of interest for icd-9}

\item{diag_ecode_col:}{indices(numeric) of diagnosis and E codes of interest}

\item{date:}{date of discharge}
}
\value{
additional variables: any_drug, any_opioid, non_heroin_opioid ,
  heroin
}
\description{
Use date to apply ICD-9_CM criteria to pre fiscal year 2016 and ICD-10-CM
criteria to the rest.
}
\examples{
to be added


}
