% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od_fatal_alchohol_tox.R
\name{od_fatal_alcohol_tox}
\alias{od_fatal_alcohol_tox}
\title{Find any Alcohol toxicity}
\usage{
od_fatal_alcohol_tox(data, underly_col, mult_col)
}
\arguments{
\item{data}{input data}

\item{underly_col}{underlying cause column index}

\item{mult_col}{multicause index}
}
\value{
alcohol_tox
}
\description{
Find any Alcohol toxicity
}
\examples{
to be added

}
