% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od_fatal_antidep.R
\name{od_fatal_antidep}
\alias{od_fatal_antidep}
\title{Find andidepressant poisoning}
\usage{
od_fatal_antidep(data, underly_col, mult_col)
}
\arguments{
\item{data}{input data}

\item{underly_col}{underlying cause column index}

\item{mult_col}{multicause index}
}
\value{
antidepressant
}
\description{
Find andidepressant poisoning
}
\examples{
to be added

}
