% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od_fatal_any_drug.R
\name{od_fatal_any_drug}
\alias{od_fatal_any_drug}
\title{Find any drug overdose.}
\usage{
od_fatal_any_drug(data, underly_col)
}
\arguments{
\item{data}{input data}

\item{underly_col}{underlying column index}
}
\value{
cdc_any_drug
}
\description{
Find any drug based on CDC definitions
}
\examples{
to be added
}
