% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od_fatal_benzo.R
\name{od_fatal_benzo}
\alias{od_fatal_benzo}
\title{Find any Benzodiazepines poisoning.}
\usage{
od_fatal_benzo(data, underly_col, var_name = "benzo")
}
\arguments{
\item{data}{input data}

\item{underly_col}{underlying column index}

\item{var_name}{propose a name for the new Benzo variable. The default is "benzo"}
}
\value{
benzo
}
\description{
Find any drug based on CDC definitions
}
\examples{
to be added
}
