% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od_fatal_drug_list.R
\name{od_fatal_drug_list}
\alias{od_fatal_drug_list}
\title{Add selected drug overdose as defined by CDC using ICD-10.}
\usage{
od_fatal_drug_list(data, underly_col, mult_col)
}
\arguments{
\item{data}{the death dataset in which to add the new drug variables}

\item{underly_col}{The column index for the with the ICD-10 underlying cause of death}

\item{mult_col}{The column indices of the multiple-cause-of-death codes}
}
\value{
The original dataset with the added drug fields.1 means drug present and 0 otherwise
}
\description{
cdc_any_drugs: Any drug overdose death,
that is an underlying cause-of-death ICD-10 codes of
X40–X44, X60–X64, X85, and Y10–Y14.
The following specific drug overdose deaths are drug overdose deaths
(defined in cdc_any_drugs) that involved the specified ICD-10 codes
in the multiple-cause-of-death fields
cdc_any_opioid: Any opioid death with
T40.0, T40.1, T40.2, T40.3, T40.4, or T40.6
cdc_heroin_t40_1: Heroin death with T40.1
cdc_opioid_t40_2: Natural and semisynthetic opioids (T40.2)
cdc_opioid_t40_3: Methadone with T40.3
cdc_opioid_t40_4: Synthetic opioids excluding methadone (T40.4)
cdc_cocaine_t40_5: Cocaine (T40.5)
cdc_stimulant_t43_6: Psychostimulants with abuse potential (T43.6).
these drug categories are not mutually exclusive
}
\examples{
Examples will be added later
}
