% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od_fatal_other.R
\name{od_fatal_other}
\alias{od_fatal_other}
\title{Find other drug poisonings.
excluding opioid, benzo, antidepressant,
psycho-stimulant, cocaine}
\usage{
od_fatal_other(data, underly_col, cond.var = cdc_any_drugs,
  var_name = "other")
}
\arguments{
\item{data}{input data}

\item{underly_col}{causes column index}

\item{cond.var}{conditional variable}

\item{var_name}{propose a name for the new other variable. The default is "other"}
}
\value{
other
}
\description{
Find any drug based on CDC definitions
}
\examples{
to be added
}
