% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od_fatal_psycho.R
\name{od_fatal_psycho}
\alias{od_fatal_psycho}
\title{Find psychotropic poisoning
antidepressant and neuroleptics}
\usage{
od_fatal_psycho(data, underly_col, mult_col)
}
\arguments{
\item{data}{input data}

\item{underly_col}{underlying cause column index}

\item{mult_col}{multicause index}
}
\value{
psycho
}
\description{
Find psychotropic poisoning
antidepressant and neuroleptics
}
\examples{
to be added

}
