% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od_fatal_sedative.R
\name{od_fatal_sedative}
\alias{od_fatal_sedative}
\title{Find any sedative toxicity}
\usage{
od_fatal_sedative(data, underly_col, mult_col)
}
\arguments{
\item{data}{input data}

\item{underly_col}{underlying cause column index}

\item{mult_col}{multicause index}
}
\value{
sedative
}
\description{
Find any sedative toxicity
}
\examples{
to be added

}
