% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od_intent_icd10cm_6.R
\name{od_intent_icd10cm_6}
\alias{od_intent_icd10cm_6}
\title{Title}
\usage{
od_intent_icd10cm_6(data, diag_ecode_col)
}
\arguments{
\item{diag_ecode_col}{}
}
\value{
intent_drugs6
}
\description{
Title
}
