% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od_intent_icd9_10cm.R
\name{od_intent_icd9_10cm}
\alias{od_intent_icd9_10cm}
\title{Find intent for ICD-10-CM.}
\usage{
od_intent_icd9_10cm(data, diag_ecode_col)
}
\arguments{
\item{data}{input data}

\item{diag_ecode_col}{column indices}
}
\value{
intent_drugs
}
\description{
Find intent for ICD-10-CM.
}
\examples{
to be added
}
