% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od_intent_icd9cm.R
\name{od_intent_icd9cm}
\alias{od_intent_icd9cm}
\title{Find intent for ICD-9-CM}
\usage{
od_intent_icd9cm(data, diag_ecode_col)
}
\arguments{
\item{data}{input data}

\item{diag_ecode_col}{column indices}
}
\value{
intent_drugs
}
\description{
Find intent for ICD-9-CM
}
\examples{
to be added
}
