% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od_non_heroin_icd10cm.R
\name{od_non_heroin_icd10cm}
\alias{od_non_heroin_icd10cm}
\title{Find non-heroin opioid ICD-10-CM.}
\usage{
od_non_heroin_icd10cm(data, diag_ecode_col)
}
\arguments{
\item{data}{input data}

\item{diag_ecode_col}{column indices}
}
\value{
non_heroin_icd10cm
}
\description{
Find non-heroin opioid ICD-10-CM.
}
\examples{
to be added
}
