% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od_ui_fatal_drug.r
\name{od_ui_fatal_drug}
\alias{od_ui_fatal_drug}
\title{Find any unintentional drug overdose.}
\usage{
od_ui_fatal_drug(data, underly_col)
}
\arguments{
\item{data}{input data}

\item{underly_col}{underlying column index}
}
\value{
cdc_any_drug
}
\description{
Find any drug based on CDC definitions
}
\examples{
to be added
}
