% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od_ui_fatal_opioid.r
\name{od_ui_fatal_opioid}
\alias{od_ui_fatal_opioid}
\title{Find any opioid.}
\usage{
od_ui_fatal_opioid(data, underly_col, mult_col, var_name = "ui_opioid")
}
\arguments{
\item{data}{input data}

\item{underly_col}{underlying column index}

\item{mult_col}{multiple causes column indices}

\item{var_name}{propose a name for the new alcohol variable. The default is "ui_opioid"}
}
\value{
cdc_any_drugs, cdc_any_opioid,
}
\description{
Using CDC definitions.
}
\examples{
to be added


}
