% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pop_15_16}
\alias{pop_15_16}
\title{Simulated Population Data.}
\format{Data frame}
\source{
simulation.
}
\usage{
pop_15_16
}
\description{
Dataset of 500 rows and 8 variables.
}
\examples{
library(dplyr)
sample_n(hosp_set, 5)

}
\keyword{datasets}
