% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{remove_stop_words}
\alias{remove_stop_words}
\title{Removes stop words from care narratives. Intended to be used internally with apply.}
\usage{
remove_stop_words(str, stop_words, remove_numbers = TRUE)
}
\arguments{
\item{str}{String of single narrative}

\item{stop_words}{Corpus of stop words to use, defaults to tidytext::stop_words}

\item{remove_numbers}{Logical flag to remove numbers from text}
}
\value{
Cleaned narrative
}
\description{
Removes stop words from care narratives. Intended to be used internally with apply.
}
