% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seacolors}
\alias{seacolors}
\title{Seahawks colors.}
\format{vector}
\source{
https://teamcolorcodes.com/seattle-seahawks-color-codes/.
}
\usage{
seacolors
}
\description{
vector of 4 hex colors.
}
\examples{
plot(1:4, pch = 15, col = seacolors)

}
\keyword{datasets}
