% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stop_words}
\alias{stop_words}
\title{Stop Words}
\format{A data frame with 1149 rows and 2 variables:
\describe{
  \item{word}{actual token}
  \item{lexicon}{lexicon the stop word comes from}
}}
\usage{
stop_words
}
\description{
Exact copy of tidytext::stop_words
}
\keyword{datasets}
