% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{tune_logistic}
\alias{tune_logistic}
\title{Train and test logistic regression classifier.}
\usage{
tune_logistic(
  training_data,
  testing_data,
  save_model = TRUE,
  model_name = paste0("opioid_logistic_model_", Sys.Date()),
  target = "opioid",
  features = "defaults"
)
}
\arguments{
\item{training_data}{formatted EMS data for model training}

\item{testing_data}{formatted EMS data for model training}

\item{save_model}{Logical flag to save model as .RDS object}

\item{model_name}{Filepath and name of model to save, .RDS extension not needed}

\item{target}{Binary (0/1) for class labels}

\item{features}{Character vector of features, "default" for recommended options}
}
\value{
List object with model and test data with predictions
\itemize{
  \item model - Fitted XGBoost object
  \item results - Test data with predictions
  \item tune_object - SVM tuning object
}
}
\description{
Train and test logistic regression classifier.
}
\examples{
tune_logistic(training_data = train, testing_data = test)
}
