% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{tune_xgboost}
\alias{tune_xgboost}
\title{Train and test XGBoost classifier.}
\usage{
tune_xgboost(
  training_data,
  testing_data,
  save_model = TRUE,
  model_name = paste0("opioid_xgboost_model_", Sys.Date()),
  target = "opioid",
  features = "defaults",
  cv_iters = 5,
  tune_iters = 50
)
}
\arguments{
\item{training_data}{formatted EMS data for model training}

\item{testing_data}{formatted EMS data for model training}

\item{save_model}{Logical flag to save model as .RDS object}

\item{model_name}{Filepath and name of model to save, .RDS extension not needed}

\item{target}{Binary (0/1) for class labels}

\item{features}{Character vector of features, "default" for recommended options}

\item{cv_iters}{Number of cross validation folds for use in tuning - must be greater than 1}

\item{tune_iters}{Number of overall tuning iterations - must be greater than 1}
}
\value{
List object with model and test data with predictions
\itemize{
  \item model - Fitted XGBoost object
  \item tune_obj - XGBoost tuning object
  \item results - Test data with predictions
}
}
\description{
Train and test XGBoost classifier.
}
\examples{
\dontrun{tune_xgboost(training_data = train, testing_data = test)}
}
