% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od_any_opioid_icd10cm.R
\name{add_any_opioid_icd10cm}
\alias{add_any_opioid_icd10cm}
\title{Title}
\usage{
add_any_opioid_icd10cm(data, diag_ecode_col)
}
\arguments{
\item{diag_ecode_col}{}
}
\description{
Title
}
