% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dea_set}
\alias{dea_set}
\title{Simulated Death Data.}
\format{Data frame}
\source{
simulation.
}
\usage{
dea_set
}
\description{
Dataset of 500 rows and 6 variables.
}
\examples{
library(dplyr)
sample_n(dea_set, 5)

}
\keyword{datasets}
