% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{icd10cm18}
\alias{icd10cm18}
\title{ICD-10-CM codes table and description v2018.}
\format{Data frame}
\source{
\url{https://www.cdc.gov/nchs/icd/icd10cm.htm}
}
\usage{
icd10cm18
}
\description{
Dataset of 71,704 rows and 2 variables.
}
\examples{
library(dplyr)
sample_n(icd10cm18, 5)

}
\keyword{datasets}
