% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od_age11.R
\name{od_age11}
\alias{od_age11}
\title{Title Make an eleven-category age grouping}
\usage{
od_age11(data, age)
}
\arguments{
\item{data:}{input data}

\item{age:}{age in years}
}
\value{
Age groups agegrp11(numeric) and age11(character)
}
\description{
Title Make an eleven-category age grouping
}
\examples{
library(tidyverse)
hosp_set \%>\% od_age11(age = age) \%>\% sample_n(5)
}
