% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od_create_cond_diag.R
\name{od_create_cond_diag}
\alias{od_create_cond_diag}
\title{Title}
\usage{
od_create_cond_diag(data, expr, colvec, ignore.case = T, perl = T, cond.var)
}
\arguments{
\item{cond.var}{}
}
\description{
Title
}
