% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od_crude_rate.R
\name{od_crude_rate}
\alias{od_crude_rate}
\title{Title}
\usage{
od_crude_rate(data, count, population, s = 1e+05, r = 1, alpha = 0.05)
}
\arguments{
\item{alpha}{}
}
\value{
crude_rate, lower_crude_rate,
and upper_crude_rate
}
\description{
Title
}
