% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od_direct_age_adjust.R
\name{od_direct_age_adjust}
\alias{od_direct_age_adjust}
\title{Title}
\usage{
od_direct_age_adjust(data, agegrp = agegrp11, count = count,
  population = population, standard_pop = NULL, s = 1e+05, r = 1,
  alpha = 0.05)
}
\arguments{
\item{alpha}{}
}
\description{
Title
}
