% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od_drug_types_icd10cm.R
\name{od_drug_types_icd10cm}
\alias{od_drug_types_icd10cm}
\title{Title}
\usage{
od_drug_types_icd10cm(data, diag_ecode_col)
}
\arguments{
\item{diag_ecode_col}{}
}
\description{
Title
}
