% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od_fatal_antidep.R
\name{od_fatal_antidep}
\alias{od_fatal_antidep}
\title{Find any Antidepressant poisoning.}
\usage{
od_fatal_antidep(data, underly_col, var_name = "antidepressant")
}
\arguments{
\item{data}{input data}

\item{underly_col}{underlying column index}

\item{var_name}{propose a name for the new Antidepressant variable. The default is "antidepressant"}
}
\value{
antidepressant
}
\description{
Find any drug based on CDC definitions
}
\examples{
to be added
}
