% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od_fatal_any_drug.R
\name{od_fatal_any_drug}
\alias{od_fatal_any_drug}
\title{Title}
\usage{
od_fatal_any_drug(data, underly_col)
}
\arguments{
\item{underly_col}{}
}
\value{
cdc_any_drug
}
\description{
Title
}
