% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od_fatal_benzo.R
\name{od_fatal_benzo}
\alias{od_fatal_benzo}
\title{Find any Benzodiazepines poisoning}
\usage{
od_fatal_benzo(data, underly_col, mult_col)
}
\arguments{
\item{data}{input data}

\item{underly_col}{underlying cause column index}

\item{mult_col}{multicause index}
}
\value{
benzo
}
\description{
Find any Benzodiazepines poisoning
}
\examples{
to be added

}
