% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od_fatal_other.R
\name{od_fatal_other}
\alias{od_fatal_other}
\title{Find other drug poisonings.
excluding opioid, sedative, antidepressant,
psycho-stimulant, cocaine, and alcohol}
\usage{
od_fatal_other(data, underly_col, mult_col)
}
\arguments{
\item{data}{input data}

\item{underly_col}{underlying cause column index}

\item{mult_col}{multicause index}
}
\value{
other_drugs
}
\description{
Find other drug poisonings.
excluding opioid, sedative, antidepressant,
psycho-stimulant, cocaine, and alcohol
}
\examples{
to be added

}
