% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od_intent_icd9cm.R
\name{od_intent_icd9cm}
\alias{od_intent_icd9cm}
\title{Title}
\usage{
od_intent_icd9cm(data, diag_ecode_col)
}
\arguments{
\item{diag_ecode_col}{}
}
\value{
intent_drugs
}
\description{
Title
}
