% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od_intent_trans.r
\name{od_intent_trans}
\alias{od_intent_trans}
\title{Title}
\usage{
od_intent_trans(data, diag_ecode_col, date)
}
\arguments{
\item{diag_ecode_col}{numeric vector of the
selected diagnosis and E codes}

\item{data:}{input data}

\item{date:}{date of discharge}
}
\value{
intent_drugs with values 1, 2, 3, 4 for
unintentional, self-harm, assault and undetermined
}
\description{
Title
}
