% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od_intent_trans.r
\name{od_intent_trans}
\alias{od_intent_trans}
\title{Title}
\usage{
od_intent_trans(data, diag_ecode_col, date)
}
\arguments{
\item{date}{}
}
\value{
intent_drugs
}
\description{
Title
}
