% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{one_hot_single_row_ems_data}
\alias{one_hot_single_row_ems_data}
\title{Creates one-hot-matrix of terms in narrative and binds to data.}
\usage{
one_hot_single_row_ems_data(
  data_in,
  col_to_one_hot = "care_narrative_adj",
  words = pull(select(filter(high_value_terms, type == "word"), token)),
  bigrams = pull(select(filter(high_value_terms, type == "bigram"), token)),
  trigrams = pull(select(filter(high_value_terms, type == "trigram"), token))
)
}
\arguments{
\item{data_in}{Formatted data to use}

\item{col_to_one_hot}{Field name for specific narrative to one-hot}

\item{words}{Words to tokenize on, defaults to high_value_terms}

\item{bigrams}{Bigrams to tokenize on, defaults to high_value_terms}

\item{trigrams}{Trigrams to tokenize on, defaults to high_value_terms}
}
\value{
Dataframe of inputted data and one-hot fields
}
\description{
Creates one-hot-matrix of terms in narrative and binds to data.
}
\examples{
one_hot_single_row_ems_data(formatted_data_frame)
}
