% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seacolors}
\alias{seacolors}
\title{Seahawks and sounders colors}
\format{vector}
\source{
https://teamcolorcodes.com/
}
\usage{
seacolors
}
\description{
vector of 7 hex colors. The first 4 are Seahawks
}
\examples{

print(seacolors)
plot(1:7, pch = 15, col = seacolors, cex = 2)

}
\keyword{datasets}
